Move js/crypto.js -> utils/. Split js/utils.js in controllers/utils/lazyLoadVue.js and utils/giLodash.js
