+++ frontend/simple/utils/crypto.js
